+++ backend/canisters/local_group_index/impl/src/jobs/mod.rs
@@ -1,14 +1,10 @@
 use crate::RuntimeState;
 
-pub mod topup_canister_pool;
-pub mod topup_canisters;
 pub mod update_controllers;
 pub mod upgrade_communities;
 pub mod upgrade_groups;
 
 pub(crate) fn start(state: &RuntimeState) {
-    topup_canister_pool::start_job_if_required(state, None);
-    topup_canisters::start_job();
     update_controllers::start_job_if_required(state);
     upgrade_communities::start_job_if_required(state);
     upgrade_groups::start_job_if_required(state);