+++ backend/canisters/local_group_index/impl/src/lifecycle/post_upgrade.rs
@@ -24,9 +24,24 @@ fn post_upgrade(args: Args) {
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
-    let mut env = init_env(data.rng_seed);
+    let mut events = Vec::new();
+    for (chat_id, group) in data.local_groups.iter() {
+        events.push(LocalGroupIndexEvent::MigrateGroup(*chat_id, group.clone()));
+    }
 
-    if !data.controllers_updated && data.update_controllers_queue.is_empty() {
+    for (community_id, community) in data.local_communities.iter() {
+        events.push(LocalGroupIndexEvent::MigrateCommunity(*community_id, community.clone()));
+    }
+
+    while let Some(canister_id) = data.canister_pool.pop() {
+        events.push(LocalGroupIndexEvent::MigrateEmptyCanister(canister_id));
+    }
+
+    let mut env = init_env(data.rng_seed);
+    let now = env.now();
+
+    if !data.controllers_updated_v2 && data.update_controllers_queue.is_empty() {
+        data.update_controllers_queue.extend(data.canister_pool.iter().copied());
         data.update_controllers_queue
             .extend(data.local_groups.iter().map(|(g, _)| CanisterId::from(*g)));
         data.update_controllers_queue
@@ -36,22 +51,18 @@ fn post_upgrade(args: Args) {
     data.local_user_index_sync_queue.set_defer_processing(true);
     data.local_user_index_sync_queue.set_state(data.local_user_index_canister_id);
 
-    let now = env.now();
-    for (chat_id, group) in data.local_groups.iter() {
-        data.local_user_index_sync_queue.push(IdempotentEnvelope {
-            created_at: now,
-            idempotency_id: env.rng().next_u64(),
-            value: LocalGroupIndexEvent::MigrateGroup(*chat_id, group.clone()),
-        })
-    }
+    data.local_user_index_sync_queue.push_many(
+        events
+            .into_iter()
+            .map(|e| IdempotentEnvelope {
+                created_at: now,
+                idempotency_id: env.rng().next_u64(),
+                value: e,
+            })
+            .collect(),
+    );
 
-    for (community_id, community) in data.local_communities.iter() {
-        data.local_user_index_sync_queue.push(IdempotentEnvelope {
-            created_at: now,
-            idempotency_id: env.rng().next_u64(),
-            value: LocalGroupIndexEvent::MigrateCommunity(*community_id, community.clone()),
-        })
-    }
+    data.local_user_index_sync_queue.set_defer_processing(false);
 
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id, data.test_mode);
     init_state(env, data, args.wasm_version);