+++ backend/canisters/online_users/client/src/lib.rs
@@ -1,5 +1,0 @@
-use canister_client::generate_candid_update_call;
-use online_users_canister::*;
-
-// Updates
-generate_candid_update_call!(mark_as_online);