+++ backend/canisters/online_users/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Increase timeout of bounded-wait calls to 5 minutes ([#7730](https://github.com/open-chat-labs/open-chat/pull/7730))
 - Include more details in failed c2c call errors ([#7749](https://github.com/open-chat-labs/open-chat/pull/7749))
+- Additionally include minutes online in the previous month ([#7884](https://github.com/open-chat-labs/open-chat/pull/7884))
 
 ## [[2.0.1687](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1687-online_users)] - 2025-04-03